Move date formatting to utils folder
