+++ PyroApp/services/alerts.service.ts
@@ -1,5 +1,6 @@
 import {LatLng} from '@charlespalmerbf/react-native-leaflet-js';
 import apiClient from './apiClient.service';
+import {formatToLongDate} from '../utils/date';
 
 const CAMERA_OPENING_ANGLE = 87;
 const CAMERA_RANGE_KM = 15;
@@ -24,7 +25,7 @@ async function getAlerts(): Promise<Alert[]> {
 
 async function getAlert(alertId: number): Promise<Alert> {
   let response = await apiClient.get(`/alerts/${alertId}/`);
-  response.data.created_at = formatDate(response.data.created_at);
+  response.data.created_at = formatToLongDate(response.data.created_at);
   return response.data;
 }
 
@@ -36,51 +37,6 @@ async function getMedia(media_id: number) {
 async function getAlertsFromEvent(event_id: number) {
   const response = await apiClient.get(`/events/${event_id}/alerts`);
   return response.data;
-}
-
-function formatDate(dateString: string): string {
-  // Création d'un objet Date à partir de la chaîne
-  const date = new Date(dateString);
-
-  // Jours de la semaine
-  const joursSemaine = [
-    'Dimanche',
-    'Lundi',
-    'Mardi',
-    'Mercredi',
-    'Jeudi',
-    'Vendredi',
-    'Samedi',
-  ];
-
-  // Mois de l'année
-  const moisAnnee = [
-    'janvier',
-    'février',
-    'mars',
-    'avril',
-    'mai',
-    'juin',
-    'juillet',
-    'août',
-    'septembre',
-    'octobre',
-    'novembre',
-    'décembre',
-  ];
-
-  // Récupération des éléments de la date
-  const jourSemaine = joursSemaine[date.getDay()];
-  const jourMois = date.getDate();
-  const mois = moisAnnee[date.getMonth()];
-  const annee = date.getFullYear();
-  const heures = ('0' + date.getHours()).slice(-2);
-  const minutes = ('0' + date.getMinutes()).slice(-2);
-
-  // Construction de la chaîne de caractères formatée
-  const dateFormatee = `${jourSemaine} ${jourMois} ${mois} ${annee} à ${heures}:${minutes}`;
-
-  return dateFormatee;
 }
 
 // Fonction pour convertir les degrés en radians
